+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## HEAD
 
+### Submission proxy: Root compatibility checking
+
+* Adds the ability for a CT client to disable root compatibile checking: https://github.com/google/certificate-transparency-go/pull/1258
+
 ## v1.1.8
 
 * Recommended Go version for development: 1.21